+++ example/src/app/select-example/select-example.component.ts
@@ -1,4 +1,9 @@
 import {Component, OnInit} from '@angular/core';
+import { FormGroup } from "@angular/forms";
+import { FormlyFieldConfig } from "@ngx-formly/core";
+import { Subject } from "rxjs";
+
+type Option = { value: string, label: string, disabled: boolean };
 
 @Component({
   selector: 'app-select-example',
@@ -13,8 +18,8 @@ import {Component, OnInit} from '@angular/core';
           <wcs-select-option [value]="3">Three</wcs-select-option>
         </wcs-select>
         <h3>Binding</h3>
-        <p>Selected values : {{binding}}</p>
-        <wcs-select placeholder="Le select" [(ngModel)]="binding" name="sel-893" multiple>
+        <p>Selected values : {{bindingCustomSelect}}</p>
+        <wcs-select placeholder="Le select" [(ngModel)]="bindingCustomSelect" name="sel-893" multiple>
           <wcs-select-option value="1">One</wcs-select-option>
           <wcs-select-option value="2">Two</wcs-select-option>
           <wcs-select-option value="3">Three</wcs-select-option>
@@ -30,15 +35,94 @@ import {Component, OnInit} from '@angular/core';
         <wcs-button class="primary" mode="stroked" (click)="onResetButtonClick()">Reset</wcs-button>
       </wcs-card-body>
     </wcs-card>
+    <h2>Exemple d'utilisation du select natif</h2>
+    <wcs-card>
+      <wcs-card-body>
+        <h3>Valeur par défaut</h3>
+        <wcs-native-select>
+          <select name="trains-type" id="trains-select-natif">
+            <option disabled hidden selected>Choisissez un train 🚅</option>
+            <option value="TGV">TGV</option>
+            <option value="TER">TER</option>
+            <option value="Intercités">Intercités</option>
+          </select>
+        </wcs-native-select>
+
+        <h3>Set value</h3>
+        <wcs-native-select>
+          <select name="trains-type" id="trains-select-natif">
+            <option disabled hidden selected>Choisissez un train 🚅</option>
+            <option value="TGV">TGV</option>
+            <option value="TER">TER</option>
+            <option value="Intercités">Intercités</option>
+          </select>
+        </wcs-native-select>
+
+        <h3>Binding</h3>
+        <p>Selected values : {{bindingNativeSelect}}</p>
+        <wcs-native-select>
+          <select name="trains-type-binding" [(ngModel)]="bindingNativeSelect" id="trains-select-binding-natif">
+            <option value="TGV">TGV</option>
+            <option value="TER">TER</option>
+            <option value="Intercités">Intercités</option>
+          </select>
+        </wcs-native-select>
+
+        <h3>Formly</h3>
+        <formly-form [formGroup]="form" [fields]="fields" [model]="model"></formly-form>
+      </wcs-card-body>
+    </wcs-card>
   `,
   styles: []
 })
-export class SelectExampleComponent {
+export class SelectExampleComponent implements OnInit {
+  private optionsFormly: Option[] = [
+    {
+      value: '1',
+      label: 'Première valeur',
+      disabled: false
+    }, {
+      value: '2',
+      label: 'Deuxième valeur',
+      disabled: true
+    }, {
+      value: '3',
+      label: 'Troisième valeur',
+      disabled: false
+    }
+  ];
+  private asynchronousOptionsSubject = new Subject<Option[]>();
+
   value = [1, 2];
-  binding: any;
+  bindingCustomSelect: any;
+  bindingNativeSelect: any = 'Intercités';
   random = this.randomIntFromInterval(1, 3);
 
+  form = new FormGroup({});
+  fields: FormlyFieldConfig[] = [
+    {
+      fieldGroup: [
+        {
+          id: 'fieldNativeSelect',
+          key: 'fieldNativeSelect',
+          type: 'native-select',
+          props: {
+            name: 'Le select natif',
+            id: 'select-formly'
+          }
+        }
+      ]
+    }
+  ];
+  model = {
+    fieldNativeSelect: ''
+  };
+
   constructor() {
+  }
+
+  ngOnInit(): void {
+    setTimeout(() => this.asynchronousOptionsSubject.next(this.optionsFormly), 1000);
   }
 
   randomIntFromInterval(min, max): number { // min and max included
@@ -50,7 +134,8 @@ export class SelectExampleComponent {
   }
 
   onResetButtonClick() {
-    this.binding = null;
+    this.bindingCustomSelect = null;
+    this.bindingNativeSelect = null;
     this.random = null;
   }
 }